Add inference info and run commands straight into model + optimize imports
