+++ python/fedml/cli/modules/login.py
@@ -1,4 +1,5 @@
 import click
+
 from fedml.cli.modules import utils
